WIP: Comment out selectors and old contexts that should not be needed anymore, added some comments, and more.
