+++ src/resources/playlists.js
@@ -5,6 +5,7 @@ export const playlistContext =
   require.context('lessonSrc/', true, /^[.][/][^/]+[/]playlists[/][^.]+[.]yml/);
 
 const playlists = {};
+// TODO: Make this a function, and change 'playlists' to 'cachedPlaylists'
 for (const key of playlistContext.keys()) {
   const [/* ignore */, course, playlist] = key.match(/^[.][/]([^/]+)[/]playlists[/]([^.]+)[.]yml$/);
   const {sortindex = 0, title = {}, lessons = []} = playlistContext(key);
@@ -19,6 +20,7 @@ const getPlaylist = (course, playlist) => (playlists[course] || {})[playlist] ||
  * @return {string[]} An array with the (file)name of all the playlists (without extension),
  *                    e.g. ['playlist1', 'playlist2']
  */
+  // TODO: Perhaps change this to getSortedPlaylists (using sortindex)?
 export const getPlaylists = (course) =>
   Object.keys(playlists[course] || {});
 
@@ -37,6 +39,7 @@ export const getPlaylistLessons = (course, playlist) =>
  * @param {string} playlist E.g. The filename of the playlist (without extension) e.g. 'playlist1'
  * @return {number} The sortindex of the playlist
  */
+  // TODO: Might not need this function, instead return playlists sorted
 export const getPlaylistSortindex = (course, playlist) =>
   getPlaylist(course, playlist).lessons || 0;
