+++ src/selectors/frontmatter.js
@@ -1,29 +1,29 @@
-import {lessonFrontmatterContext} from '../contexts';
-
-// import {readmeContext} from '../contexts';
-// import {getLessonData} from '../util';
-
+// import {lessonFrontmatterContext} from '../contexts';
+//
+// // import {readmeContext} from '../contexts';
+// // import {getLessonData} from '../util';
+//
+// // const getLessonFrontmatter = (state, {course, lesson, file}) => {
+// //   if (!file) { return {}; }
+// //   const lessons = getLessonData();
+// //   const lessonPath = `./${course}/${lesson}/${file}.md`;
+// //   const isReadme = /README(_[a-z]{2})?/.test(file);
+// //   // NOTE: Frontmatter for README does not contain tags anymore.
+// //   return isReadme ? readmeContext(lessonPath).frontmatter : (lessons[lessonPath] || {});
+// // };
+//
 // const getLessonFrontmatter = (state, {course, lesson, file}) => {
 //   if (!file) { return {}; }
-//   const lessons = getLessonData();
 //   const lessonPath = `./${course}/${lesson}/${file}.md`;
-//   const isReadme = /README(_[a-z]{2})?/.test(file);
-//   // NOTE: Frontmatter for README does not contain tags anymore.
-//   return isReadme ? readmeContext(lessonPath).frontmatter : (lessons[lessonPath] || {});
+//   return lessonFrontmatterContext(lessonPath) || {};
 // };
-
-const getLessonFrontmatter = (state, {course, lesson, file}) => {
-  if (!file) { return {}; }
-  const lessonPath = `./${course}/${lesson}/${file}.md`;
-  return lessonFrontmatterContext(lessonPath) || {};
-};
-
-export const getTitle = (state, params) => getLessonFrontmatter(state, params).title || '';
-
-export const getLevel = (state, params) => getLessonFrontmatter(state, params).level || 0;
-
-export const getTags = (state, params) => getLessonFrontmatter(state, params).tags || {};
-
-export const getAuthorName = (state, params) => getLessonFrontmatter(state, params).author || '';
-
-export const getTranslatorName = (state, params) => getLessonFrontmatter(state, params).translator || '';
+//
+// export const getTitle = (state, params) => getLessonFrontmatter(state, params).title || '';
+//
+// export const getLevel = (state, params) => getLessonFrontmatter(state, params).level || 0;
+//
+// export const getTags = (state, params) => getLessonFrontmatter(state, params).tags || {};
+//
+// export const getAuthorName = (state, params) => getLessonFrontmatter(state, params).author || '';
+//
+// export const getTranslatorName = (state, params) => getLessonFrontmatter(state, params).translator || '';
